fe: Add new Javascript implementation
This patch replaces the existing Javascript implementation with a new
one. It's fairly minimal in feel but adds a bunch of new features:

* Enable draggable split between list pane and message pane
* Add keyboard shortcuts to select next/previous message ("j" and "k")
* Maintain scroll position when switching between views
* Minimize the amount of content rendered by Javascript (C is faster)

This patch also improves the styling of the pagination controls.

In order to make implementing these things easier, this patch also
brings in two additional external files: a CSS "normalize"
implementation to make it easier to control the look & feel
cross-browser, and the JQuery library which makes it easier to write
cross-browser Javascript.