Add support for arbitrary include files in header and footer
This allows us to include multiple CSS and Javascript files, which is
much more preferable to do for latency reasons then including other CSS
and Javascript files from one (because of the multiple RTT latencies
involved when a file must be downloaded and parsed before another
required file can be fetched by the browser).

This makes it much easier to add helpful libraries like JQuery.

Also moved the Javascript include down to the bottom of the body,
instead of the header. For info on why, see
https://developer.yahoo.com/blogs/ydn/high-performance-sites-rule-6-move-scripts-bottom-7200.html

TODO: Add option to only reopen the include file FDs when the server is
restarted, which will result in better performance due to
filesystem-level caching of the data as well as get rid of stat() calls
per request.